+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.3-dev
 
 - Do not override attributes and styles when the value matches the previous one (and the DOM may be different).
+- Relax DOM reuse restrictions on style properties.
 
 ## 0.3.2+1