+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Do not override attributes and styles when the value matches the previous one (and the DOM may be different).
 - Relax DOM reuse restrictions on style properties.
+- Relax DOM reuse restrictions on event listeners.
 
 ## 0.3.2+1